requirements.txt
Flask
pandas
selenium
webdriver-manager
beautifulsoup4
Flask-Cors
beautifulsoup4
python-dotenv
requests
flask-pymongo 
flask-cors 
pyjwt 
werkzeug

